Update food_page.php
javascript = food_page.js ... Unsure how you want to proceed with comments, adding to cart and so forth. I tried to find the method checkProductId() which the product_page.php relies on to get comments, reviews, add to cart. I thought about creating a new ingredient anytime it couldn't find a product id, but i think that may over populate the database and cause the ajax_listing.php to return all the ingredients that are searched for...